@@ -1,26 +1,27 @@
 {
   "name": "my-app",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^6.0.0-beta2",
-    "@fortawesome/fontawesome-svg-core": "^1.3.0-beta2",
-    "@fortawesome/free-regular-svg-icons": "^6.0.0-beta2",
-    "@fortawesome/react-fontawesome": "^0.1.16",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
     "babel-plugin-macros": "^3.1.0",
-    "bootstrap": "^5.1.1",
-    "node-sass": "^7.0.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-player": "^2.10.0",
-    "react-popper": "^2.2.5",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
+    "bootstrap": "^5.2.2",
+    "node-sass": "^8.0.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.6.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.6.0",
+    "react-player": "^2.11.0",
+    "react-popper": "^2.3.0",
+    "react-scripts": "^5.0.1",
+    "react-single-page-navigation": "^1.0.1",
+    "react-svg": "^15.1.12",
+    "web-vitals": "^3.0.4"
   },
   "scripts": {
     "start": "react-app-rewired start",
@@ -49,6 +50,7 @@
   "devDependencies": {
     "csp-html-webpack-plugin": "^5.1.0",
     "customize-cra": "^1.0.0",
-    "react-app-rewired": "^2.1.8"
+    "react-app-rewired": "^2.2.1",
+    "react-router-dom": "^6.4.3"
   }
 }
