@@ -1,27 +1,36 @@
 import Container from "react-bootstrap/Container"
 import Row from "react-bootstrap/Row"
 import Col from "react-bootstrap/Col"
-import MyButton from "../Email-Button"
+import Button from "react-bootstrap/Button"
+
+import {button} from "../../data.js"
+
 import "./index.css"
 
-export default function MyHeader() {
+const MyHeader = (props) => {
+    const {title, paragraph} = props
     return (
     <>
         <section id="contact" className="viewport">
             <Container>
                 <Row>
-                    <Col> <h2 id="contact-title" className="contact-text">Get in Contact</h2> </Col>
+                    <Col> <h2 id="title" className="text">{title}</h2> </Col>
                 </Row>
             <br />
                 <Row>
                     <Col>
-                        <p className="contact-text">Get in Contact with Me. Click on this Button to Send an Email to Me,
-                        or reach out to me on the listed platforms below.</p>
+                        <p className="text">{paragraph}</p>
                     </Col>
                 </Row>
-                <MyButton />
+                <Row id="email-container">
+                <Col>
+                    <Button id={`${button[4].id}`} variant={`${button[4].variantType}`} href={`${button[4].hLink}`} size={`${button[4].size}`}> {`${button[4].name}`} </Button>
+                </Col>
+            </Row>
             </Container>
         </section>
     </> 
     );
 }
+
+export default MyHeader
